+++ CHANGES.md
@@ -2,3 +2,4 @@
 * initial pass
 * attempt to make both selective and inclusive to only primary menu
 * attempt testing with variety of themes
+* try to make compatible with block navigation menu